requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
pydantic
python-multipart==0.0.9
aiohttp==3.9.3
sqlalchemy
typing-extensions>=4.8.0
openai==1.77.0
faiss-cpu
numpy